+++ MatomoTracker.php
@@ -28,6 +28,7 @@ class MatomoTracker
      * MatomoTracker::$URL = 'http://yourwebsite.org/matomo/';
      *
      * @var string
+     * @deprecated
      */
     static public $URL = '';
 
@@ -213,6 +214,8 @@ class MatomoTracker
 
     private $requestMethod = null;
 
+    private $apiUrl = '';
+
     /**
      * Builds a MatomoTracker object, used to track visits, pages and Goal conversions
      * for a specific website, by using the Matomo Tracking API.
@@ -234,10 +237,12 @@ class MatomoTracker
             !empty($_SERVER['HTTP_SEC_CH_UA_PLATFORM']) ? $_SERVER['HTTP_SEC_CH_UA_PLATFORM'] : '',
             !empty($_SERVER['HTTP_SEC_CH_UA_PLATFORM_VERSION']) ? $_SERVER['HTTP_SEC_CH_UA_PLATFORM_VERSION'] : '',
             !empty($_SERVER['HTTP_SEC_CH_UA_FULL_VERSION_LIST']) ? $_SERVER['HTTP_SEC_CH_UA_FULL_VERSION_LIST'] : '',
-            !empty($_SERVER['HTTP_SEC_CH_UA_FULL_VERSION']) ? $_SERVER['HTTP_SEC_CH_UA_FULL_VERSION'] : ''
+            !empty($_SERVER['HTTP_SEC_CH_UA_FULL_VERSION']) ? $_SERVER['HTTP_SEC_CH_UA_FULL_VERSION'] : '',
+            !empty($_SERVER['HTTP_SEC_CH_UA_FORM_FACTORS']) ? $_SERVER['HTTP_SEC_CH_UA_FORM_FACTORS'] : ''
         );
         if (!empty($apiUrl)) {
             self::$URL = $apiUrl;
+            $this->apiUrl = $apiUrl;
         }
 
         $this->setNewVisitorId();
@@ -251,6 +256,7 @@ class MatomoTracker
     public function setApiUrl(string $url): void
     {
         self::$URL = $url;
+        $this->apiUrl = $url;
     }
 
     /**
@@ -604,6 +610,8 @@ class MatomoTracker
      *      or an array containing all brands with the structure
      *      [['brand' => 'Chrome', 'version' => '10.0.2'], ['brand' => '...]
      * @param string $uaFullVersion  Value of the header 'HTTP_SEC_CH_UA_FULL_VERSION'
+     * @param string|array<string> $formFactors  Value of the header 'HTTP_SEC_CH_UA_FORM_FACTORS'
+     *      or an array containing all form factors with structure ["Desktop", "XR"]
      *
      * @return $this
      */
@@ -612,7 +620,8 @@ class MatomoTracker
         string $platform = '',
         string $platformVersion = '',
         $fullVersionList = '',
-        string $uaFullVersion = ''
+        string $uaFullVersion = '',
+        $formFactors = ''
     ) {
         if (is_string($fullVersionList)) {
             $reg  = '/^"([^"]+)"; ?v="([^"]+)"(?:, )?/';
@@ -628,12 +637,25 @@ class MatomoTracker
             $fullVersionList = [];
         }
 
+        if (is_string($formFactors)) {
+            $formFactors = explode(',', $formFactors);
+            $formFactors = array_filter(array_map(
+                function ($item) {
+                    return trim($item, '" ');
+                },
+                $formFactors
+            ));
+        } elseif (!is_array($formFactors)) {
+            $formFactors = [];
+        }
+
         $this->clientHints = array_filter([
             'model' => $model,
             'platform' => $platform,
             'platformVersion' => $platformVersion,
             'uaFullVersion' => $uaFullVersion,
             'fullVersionList' => $fullVersionList,
+            'formFactors' => $formFactors,
         ]);
 
         return $this;
@@ -2215,23 +2237,29 @@ didn't change any existing VisitorId value */
 
     /**
      * Returns the base URL for the Matomo server.
+     *
+     * @throws Exception
      */
     protected function getBaseUrl(): string
     {
-        if (empty(self::$URL)) {
+        $apiUrl = $this->apiUrl === ''
+            ? self::$URL
+            : $this->apiUrl;
+
+        if ($apiUrl === '') {
             throw new Exception(
                 'You must first set the Matomo Tracker URL by calling
                  MatomoTracker::$URL = \'http://your-website.org/matomo/\';'
             );
         }
-        if (strpos(self::$URL, '/matomo.php') === false
-            && strpos(self::$URL, '/proxy-matomo.php') === false
+        if (strpos($apiUrl, '/matomo.php') === false
+            && strpos($apiUrl, '/proxy-matomo.php') === false
         ) {
-            self::$URL = rtrim(self::$URL, '/');
-            self::$URL .= '/matomo.php';
+            $apiUrl = rtrim($apiUrl, '/');
+            $apiUrl .= '/matomo.php';
         }
 
-        return self::$URL;
+        return $apiUrl;
     }
 
     /**